refactor(#5): split compliance bulk upload into hook and panels
Extract types, helpers, state hook, groups panel, and footer so the dialog stays a thin orchestrator.
